Adds explanation about difference README and hosts
hosts file described in the README is not exactly the same. This is
because, if we don't specify a user, ansible will use it's defaults
(remote_user in ~/.ansible.cfg) and the playbook potentially won't work
for some people.
This forces ssh'ing as root, and fixes this potential problem.
Fixes #25