Update ENV variable
Following Ansible 2.7 preco [DEPRECATION WARNING]: ANSIBLE_HOSTS option, The variable is misleading as it can be a list of hosts and/or paths to inventory 
sources , use ANSIBLE_INVENTORY instead. This feature will be removed in version 2.8. Deprecation warnings can be disabled by 
setting deprecation_warnings=False in ansible.cfg.